+++ lcviz/tests/test_plugin_flux_column.py
@@ -4,7 +4,7 @@ from numpy.testing import assert_allclose
 def test_docs_snippets(helper, light_curve_like_kepler_quarter):
     lcviz, lc = helper, light_curve_like_kepler_quarter
 
-    lcviz.load_data(lc)
+    lcviz.load(lc)
     # lcviz.show()
 
     flux_col = lcviz.plugins['Flux Column']