+++ lcviz/tests/test_plugin_frequency_analysis.py
@@ -6,7 +6,7 @@ from lightkurve.periodogram import LombScarglePeriodogram, BoxLeastSquaresPeriod
 def test_docs_snippets(helper, light_curve_like_kepler_quarter):
     lcviz, lc = helper, light_curve_like_kepler_quarter
 
-    lcviz.load_data(lc)
+    lcviz.load(lc)
     # lcviz.show()
 
     freq = lcviz.plugins['Frequency Analysis']