+++ lcviz/tests/test_plugin_markers.py
@@ -23,7 +23,7 @@ def _assert_dict_allclose(dict1, dict2):
 def test_docs_snippets(helper, light_curve_like_kepler_quarter):
     lcviz, lc = helper, light_curve_like_kepler_quarter
 
-    lcviz.load_data(lc)
+    lcviz.load(lc)
     # lcviz.show()
 
     markers = lcviz.plugins['Markers']
@@ -141,7 +141,7 @@ def test_tpf_markers(helper, light_curve_like_kepler_quarter):
     label_mouseover = mp._obj.coords_info
     mp.open_in_tray()
 
-    assert abs(helper.plugins['Time Selector'].value - 47.006888) < 1e-4
+    assert abs(helper.plugins['Time Selector'].value - 46.998069) < 1e-4
 
     # test event in image (TPF) viewer
     iv = helper.viewers['image']._obj
@@ -149,18 +149,19 @@ def test_tpf_markers(helper, light_curve_like_kepler_quarter):
                                         {'event': 'mousemove',
                                          'domain': {'x': 0, 'y': 0}})
 
-    assert label_mouseover.as_text() == ('Pixel x=00000.0 y=00000.0 Value +1.28035e+01 electron / s',  # noqa
-                                         'Time 47.00689 d',
+    # print(label_mouseover.as_text())
+    assert label_mouseover.as_text() == ('Pixel x=00000.0 y=00000.0 Value +1.00643e+01 electron / s',  # noqa
+                                         'Time 46.99807 d',
                                          '')
 
-    print(label_mouseover.as_dict())
-    _assert_dict_allclose(label_mouseover.as_dict(), {'data_label': 'KIC 1429092[TPF]',
-                                                      'time': 47.00689,
+    # print(label_mouseover.as_dict())
+    _assert_dict_allclose(label_mouseover.as_dict(), {'data_label': 'KIC 1429092 [TPF]',
+                                                      'time': 46.99807,
                                                       'time:unit': 'd',
                                                       'pixel': (0.0, 0.0),
                                                       'axes_x': 0,
                                                       'axes_x:unit': 'pix',
                                                       'axes_y': 0,
                                                       'axes_y:unit': 'pix',
-                                                      'value': 12.803529,
+                                                      'value': 10.064313,
                                                       'value:unit': 'electron / s'})